apply node.js feedback
